+++ test/coverage/test_phase.py
@@ -1004,6 +1004,43 @@ def test_underspecified_path_parameters(ctx, cli, snapshot_cli, openapi3_base_ur
     )
 
 
+def test_path_parameters_arent_missing(ctx, cli, snapshot_cli, openapi3_base_url):
+    # When `--mode=negative`, still generate path parameters if they can't be negated
+    schema_path = ctx.openapi.write_schema(
+        {
+            "/organizations/{organization_id}/": {
+                "get": {
+                    "parameters": [
+                        {
+                            "name": "organization_id",
+                            "in": "path",
+                            "required": True,
+                            "schema": {"type": "string"},
+                        },
+                        {
+                            "name": "q",
+                            "in": "query",
+                            "required": True,
+                            "schema": {"type": "integer", "minimum": 10},
+                        },
+                    ],
+                    "responses": {"200": {"description": "Successful Response"}},
+                }
+            }
+        }
+    )
+    assert (
+        cli.run(
+            str(schema_path),
+            f"--url={openapi3_base_url}",
+            "--checks=not_a_server_error",
+            "--phases=coverage",
+            "--mode=negative",
+        )
+        == snapshot_cli
+    )
+
+
 def test_path_parameter_dots(ctx):
     schema = build_schema(
         ctx,