+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 - Properly search for `examples` in Open API 2.0 responses.
 - Coverage phase missing some examples for request bodies.
 - Accept booleans for parameter schemas in Open API 3.1.
-- Fixed false positives in `negative_data_rejection` check for array of strings query parameters. [#3056](https://github.com/schemathesis/schemathesis/issues/3056)
+- False positives in `negative_data_rejection` check for array of strings query parameters. [#3056](https://github.com/schemathesis/schemathesis/issues/3056)
+- False positive report about missing path parameter during the coverage phase with `--mode=negative`.
 
 ### :wrench: Changed